refactor into grid layouts
