+++ front/src/app/__tests__/App.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '../../app/App';
+import App from '../App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
