+++ front/src/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 
-import App from './components/app/App';
+import App from './app/App';
 import reducers from './reducers/toolboxReducer';
 import * as serviceWorker from './serviceWorker';
